requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6  # 用于处理文件上传

# File Processing
python-docx>=1.0.0  # 处理docx文件
PyPDF2>=3.0.0      # 处理PDF文件
openpyxl>=3.1.2    # 处理Excel文件

# Data Processing
pandas>=2.1.3      # 数据处理和CSV操作

# LLM API
google-generativeai>=0.3.0  # Gemini API

# Testing
pytest>=7.4.3

# 开发环境依赖
pytest-cov>=4.1.0  # 代码覆盖率测试
pytest-mock>=3.14.0  # 测试模拟
pytest-asyncio>=0.21.0  # 异步测试
